refactor json module without adding tests
