+++ commons/commons-model-extended-serialization/build.gradle.kts
@@ -9,6 +9,7 @@ import TestDependencies.BaseTestRuntimeDependencies
 dependencies {
     api(project(":commons:commons-model-extended"))
     api(project(":commons:commons-serialization-util"))
+    api(project(":commons:commons-json-util"))
 
     implementation(Arrow)
     implementation(JacksonKotlin)
