+++ settings.gradle
@@ -1,6 +1,7 @@
 rootProject.name = 'biostudies-backend-services'
 
 include("commons:commons-util")
+include("commons:commons-json-util")
 include("commons:commons-bio")
 include("commons:commons-http")
 include("commons:commons-serialization")
